merge dev to master

+++ app/src/main/java/com/github/malitsplus/shizurunotes/utils/FileUtils.java
@@ -1,21 +1,11 @@
-package com.github.malitsplus.shizurunotes.common;
-
-import android.os.Build;
-
-import androidx.annotation.NonNull;
-
-import com.github.malitsplus.shizurunotes.data.Chara;
+package com.github.malitsplus.shizurunotes.utils;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.lang.reflect.Field;
-import java.lang.reflect.Type;
-import java.text.DecimalFormat;
-import java.util.ArrayList;
 
-public class Utils {
+public class FileUtils {
     /***
      * 单纯的复制文件
      * @param srcPath 源文件路径
@@ -66,53 +56,10 @@ public class Utils {
     }
 
     public static void checkFileAndDeleteIfExists(File file){
-        if (file.exists())
-            deleteFile(file);
+        if (file.exists()) deleteFile(file);
     }
 
-    /***
-     * 以逗号分隔字符串
-     * @param list
-     * @return
-     */
-    public static String splitIntegerWithComma(ArrayList<Integer> list){
-        if(list.isEmpty())
-            return null;
+    public static void saveToInternalFile(String path, String file) {
 
-        StringBuilder sb = new StringBuilder();
-        for(Integer item : list){
-            sb.append(item).append(",");
-        }
-        sb.deleteCharAt(sb.lastIndexOf(","));
-        return sb.toString();
-    }
-
-
-    public static Object getValueFromObject(@NonNull Object object, @NonNull String fieldName) {
-        Field field;
-        try {
-            field = object.getClass().getDeclaredField(fieldName);
-            field.setAccessible(true);
-            return field.get(object);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    public static boolean isAtLeastVersion(int version) {
-        return Build.VERSION.SDK_INT >= version;
-    }
-
-
-    private static DecimalFormat format = new DecimalFormat("#");
-    public static String roundDownDouble(double value){
-        return format.format(Math.floor(value));
-    }
-    public static String roundUpDouble(double value){
-        return format.format(Math.ceil(value));
-    }
-    public static String roundDouble(double value){
-        return format.format(Math.round(value));
     }
 }
